+++ app/pb_kits/playbook/pb_date_range_stacked/docs/_date_range_stacked_dark.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { DateRangeStacked } from '../../'
+
+const DateRangeStackedDark = () => (
+  <div>
+    <DateRangeStacked
+        dark
+        endDate={new Date('20 Mar 2020')}
+        startDate={new Date('18 Jun 2019')}
+    />
+  </div>
+)
+
+export default DateRangeStackedDark