+++ app/pb_kits/playbook/pb_popover/popover.rb
@@ -11,12 +11,18 @@ module Playbook
       prop :tooltip_id
       prop :offset, type: Playbook::Props::Boolean, default: false
 
-      def popover_offset
-        offset ? "0, 8" : "0, 0"
-      end
-
       def classname
         generate_classname("pb_popover_kit")
+      end
+
+      def data
+        Hash(values[:data]).merge(
+          pb_popover_kit: true,
+          pb_popover_position: position,
+          pb_popover_trigger_element_id: trigger_element_id,
+          pb_popover_tooltip_id: tooltip_id,
+          pb_popover_offset: offset,
+        )
       end
     end
   end